meals: the backfill's seasoning test defeated its own safety rule
The converter only writes a recipe when every line yields an amount or is recognisably
seasoning, because a quantity:null row is EXCLUDED from the macro total -- a line whose
amount we failed to read would silently delete real calories rather than merely soften
the number.

That test was a substring match over the whole line, and /\bpepper\b/ matched
"Onion, bell pepper, marinara". A line carrying marinara and onion was therefore
classified as seasoning, and Turkey Pasta would have been converted with those calories
dropped -- the exact failure the all-or-nothing rule exists to prevent, defeated by its
own allowlist. Caught by auditing every quantity:null row the dry run produced before
writing any of them.

The unit of judgement is now the COMPONENT, not the line: a line is negligible only when
every comma/&/plus-separated part is individually on the list. Turkey Pasta joins the
blocked set, which goes from 5 to 6; conversions go 67 -> 66.

The conversion also moves out of scripts/ into src/lib/nutrition/recipe-backfill.ts. The
CI node job runs `node --test` with NO npm install -- deliberately, so unit tests stay
pure logic -- so a test importing the script failed to resolve @supabase/supabase-js
through it. Parsing now lives in lib where tests can reach it; the script keeps only IO.
That is also why main() no longer runs on import: the module exports its converters for
exactly the offline audit that caught this bug.

Six tests pin the behaviour, including the bell-pepper case by name.
